fix build on darwin
